refactor: 자동차경주(리팩토링) - 패키지 분리
