+++ src/main/java/racingCar/controller/RacingGameMain.java
@@ -1,13 +1,17 @@
-package racingCar;
+package racingCar.controller;
 
 import java.util.List;
+import racingCar.domain.Car;
+import racingCar.domain.OverCarNameLengthException;
+import racingCar.domain.RacingGame;
+import racingCar.view.InputView;
+import racingCar.view.OutputView;
 
 public class RacingGameMain {
 
   public static void main(String[] args) {
     InputView inputView = new InputView();
     OutputView outputView = new OutputView();
-    Referee referee = new Referee();
 
     String[] carNames = inputView.inputCarNames();
     int tryCount = inputView.inputTryCount();
