+++ src/main/java/racingCar/domain/MovableStrategy.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.domain;
 
 public interface MovableStrategy {
