+++ src/main/java/racingCar/domain/OverCarNameLengthException.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.domain;
 
 public class OverCarNameLengthException extends RuntimeException {
