+++ src/main/java/racingCar/domain/RandomMovable.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.domain;
 
 import java.util.Random;
