+++ src/main/java/racingCar/domain/Referee.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
